+++ src/lib/components/TrainingCourseCard.svelte
@@ -18,7 +18,7 @@
             <h4 class="subtitle">{subtitle}</h4>
             {/if}
         </div>
-        <a href="/inquire?id={id}" class="book" target="_blank">Inquire Now</a>
+        <a href="/train?id={id}" class="book" target="_blank">Inquire Now</a>
     </div>
 
     {#if heading}
