+++ src/lib/style/layout.css
@@ -39,6 +39,33 @@
     grid-template-columns: repeat(4, 1fr);
     gap: 2rem;
 }
+.gap-1 {
+    gap: 0.25rem;
+}
+.gap-2 {
+    gap: 0.5rem;
+}
+.gap-4 {
+    gap: 1rem;
+}
+.gap-6 {
+    gap: 1.5rem;
+}
+.gap-8 {
+    gap: 2rem;
+}
+.gap-10 {
+    gap: 2.5rem;
+}
+.gap-12 {
+    gap: 3rem;
+}
+.gap-14 {
+    gap: 3.5rem;
+}
+.gap-16 {
+    gap: 4rem;
+}
 .align-center {
     align-items: center;
 }
@@ -47,6 +74,9 @@
 }
 .justify-between {
     justify-content: space-between;
+}
+.justify-center {
+    justify-content: center;
 }
 .pos-absolute {
     position: absolute;
